reorganize files and enum param deref properly
